Adding student self registration

+++ database/migrations/2025_01_28_031310_add_session_programme_id_to_students_table.php
@@ -15,6 +15,8 @@ return new class extends Migration
             $table->unsignedBigInteger('session_programme_id')->after('user_id')->nullable();
             $table->foreign('session_programme_id')->references('id')->on('session_programmes')->onDelete('cascade');
         });
+
+        
     }
 
     /**
@@ -23,8 +25,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('students', function (Blueprint $table) {
-            $table->dropForeign(['sessionProgramme_id']);
-            $table->dropColumn('sessionProgramme_id');
+            if (Schema::hasColumn('students', 'sessionprogramme_id')) {
+                $table->dropForeign(['sessionprogramme_id']);
+                $table->dropColumn('sessionProgramme_id');
+            }
         });
     }
 };
